Restructured the project into 9-10 microservices
