+++ ratings/app.py
@@ -3,7 +3,7 @@ from pymongo import MongoClient, ReturnDocument
 from bson import ObjectId
 from bson.json_util import dumps, loads
 
-from restaurants.app import jsonify_with_objectid
+from Restaurants.app import jsonify_with_objectid
 
 # Connect to MongoDB
 client = MongoClient("localhost", 27017)
@@ -17,20 +17,6 @@ def add_Ratings(restuarant,rating):
    return (db['restaurant'].db.find_one_and_update({'name':restuarant},
                         { '$set': { "Rating" : rating} },
                         return_document = ReturnDocument.AFTER))
-
-
-@app.route("/FiveStarList", methods=["GET"])
-def five_star_list():
-    # pull all 5 star restaurant from table and display
-    top_five_restaurants = dict(db['restaurant'].find({"Rating": "5"}))
-    return jsonify_with_objectid(top_five_restaurants)
-
-
-@app.route("/TopThreeRestaurants", methods=["GET"])
-def top_three():
-    # pull first 3 restaurants from table and display
-    top_three_restaurants = dict(db['restaurant'].find().sort('Rating',-1).limit(3))
-    return jsonify_with_objectid(top_three_restaurants)
 
 
 if __name__ == "__main__":
