+++ restaurants/app.py
@@ -25,30 +25,6 @@ def get_restaurants():
     restaurants = list(db['restaurants'].find({}))
     return jsonify_with_objectid_array(restaurants)
 
-# API Endpoint for listing all the restaurants on the basis of 
-# Name.
-@app.route('/api/restaurants/<string:Name>', methods=['GET'])
-def get_restaurants_seats_available(Name):
-    global restaurant
-    try:
-        restaurant = db['restaurants'].find_one({"Name" : Name})
-    except:
-        print("Please try another restaurant.")
-    return jsonify_with_objectid(restaurant)
-        print("Error while retrieving a restaurant")
-    if restaurant ==None:
-         return jsonify({"message": f"Restaurant is not available. Please try another restaurant"})
-    else:
-        return jsonify_with_objectid(restaurant)
-
-
-
-# API Endpoint for listing all the restaurants on the basis of 
-# number of seats required and available.
-# @app.route('/api/restaurants/<Seats_Required>', methods=['GET'])
-# def get_restaurants_seats_available(Seats_Required):
-#     restaurant = db['restaurants'].find({"Available_Seats" : {"$et": Seats_Required, "$gt": Seats_Required}})
-#     return jsonify_with_objectid(restaurant)
 
 if __name__ == '__main__':
     app.run(debug=True)
